requirements.txt
Django==1.10.4
django-autoslug-iplweb>==1.9.4.dev0
django-extra-views>=0.7.1
django-form-utils>=1.0.3
django-gravatar2>=1.4.0
django-mptt>=0.8.4
django-widget-tweaks>=1.4.1
jsonfield>=1.0.3
Pillow>=3.2.0


